refactor(integration-tests): extract scaffold into dedicated module
